Improve prometheus support + add grafana and default dashboard
For Kubernetes we can use prometheus service discovery. Make prometheus
easier to enable/disable as well as pushgateway settings.

Add prometheus labels to metrics for app name and environment to make it
easier to dashboard

Add grafana with a default spring dashboard.

Remove `kafka-headless` service from client deployment configs, is not
needed for clients and is intended for self-clustering.

Signed-off-by: Paul Czarkowski <username.taken@gmail.com>